ajax for ordering

+++ client/src/pages/CartPage.jsx
@@ -7,11 +7,16 @@ import Product from "../components/Product";
 import cart from "../assets/images/empty-cart.svg";
 import "../assets/styles/cartpage.scss";
 import Modal from "../components/Modal";
+import axios from "axios";
 
 export default function CartPage() {
   const { cartedProducts } = useSelector((state) => state.productReducer);
   const dispatch = useDispatch();
+
   const [isOpenedModal, setIsOpenedModal] = useState(false);
+  const [nameForm, setNameForm] = useState("");
+  const [emailForm, setEmailForm] = useState("");
+  const [phoneForm, setPhoneForm] = useState("");
 
   const returnCartedProductsCost = (products) => {
     let sum = 0;
@@ -26,7 +31,6 @@ export default function CartPage() {
       <main className="cartpage">
         <div className="container">
           <aside>
-
             <div className="about__order">
               <p>Your goods({cartedProducts.length})</p>
               <p>{returnCartedProductsCost(cartedProducts)}$</p>
@@ -61,17 +65,45 @@ export default function CartPage() {
         <Modal isOpened={isOpenedModal} setIsOpenedModal={setIsOpenedModal}>
           <form
             onSubmit={(e) => {
-              e.preventDefault();
-              setIsOpenedModal(false);
-              dispatch(resetCart())
-              window.localStorage.clear()
+              if (nameForm !== "" && phoneForm !== "" && emailForm !== "") {
+                e.preventDefault();
+                setIsOpenedModal(false);
+
+                dispatch(resetCart());
+                window.localStorage.clear();
+
+                axios.post("http://localhost:3001/orders", {
+                  email: emailForm,
+                  phone: phoneForm,
+                  name: nameForm,
+                  productsId: cartedProducts.map((el) => el.id),
+                });
+              } else {
+                e.preventDefault()
+                e.target.querySelector('.order__title').innerHTML = 'Fill all fields!'
+              }
             }}
             className="order__data"
           >
             <h2 className="order__title">Сheck your order</h2>
-            <input type="text" placeholder="Your name" />
-            <input type="text" placeholder="Your phone number" />
-            <input type="email" placeholder="Your email" />
+            <input
+              value={nameForm}
+              onChange={(e) => setNameForm(e.target.value)}
+              type="text"
+              placeholder="Your name"
+            />
+            <input
+              value={phoneForm}
+              onChange={(e) => setPhoneForm(e.target.value)}
+              type="text"
+              placeholder="Your phone number"
+            />
+            <input
+              value={emailForm}
+              onChange={(e) => setEmailForm(e.target.value)}
+              type="email"
+              placeholder="Your email"
+            />
             <button type="submit">Order</button>
           </form>
         </Modal>
